+++ lua/vulkanite/groups/integrations/bufferline.lua
@@ -5,7 +5,7 @@ M.url = "https://github.com/akinsho/bufferline.nvim"
 function M.get(colors, opts)
   return {
     -- Bufferline selected buffer name.
-    BufferLineBufferSelected = { fg = colors.accent, bold = true, italic = true },
+    BufferLineBufferSelected = { fg = colors.accent_secondary, bold = true, italic = true },
   }
 end